JMM 02/04/20 - FO4 UserInformation - Update layout changing age for birthday and adding name and surname
