+++ app/src/main/java/com/example/happyroutine/ui/activity/log_in.kt
@@ -1,16 +1,15 @@
-package com.example.happyroutine
+package com.example.happyroutine.ui.activity
 import android.content.Context
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.util.Patterns
 import android.view.View
 import android.widget.Toast
+import com.example.happyroutine.R
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
 import kotlinx.android.synthetic.main.activity_log_in.*
-import kotlin.math.log
 
 
 class log_in : AppCompatActivity() {
