+++ app/src/main/java/com/example/happyroutine/ui/activity/settings.kt
@@ -1,9 +1,10 @@
-package com.example.happyroutine
+package com.example.happyroutine.ui.activity
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
+import com.example.happyroutine.R
 
 class settings : AppCompatActivity() {
