+++ app/src/main/java/com/example/happyroutine/ui/activity/social.kt
@@ -1,7 +1,8 @@
-package com.example.happyroutine
+package com.example.happyroutine.ui.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.happyroutine.R
 
 class social : AppCompatActivity() {
