+++ app/src/main/java/com/example/happyroutine/ui/dialog/DatePickerFragment.kt
@@ -1,0 +1,30 @@
+package com.example.happyroutine.ui.dialog
+
+import android.app.DatePickerDialog
+import android.app.Dialog
+import android.os.Bundle
+import androidx.fragment.app.DialogFragment
+import java.util.*
+
+class DatePickerFragment : DialogFragment() {
+
+    private var listener: DatePickerDialog.OnDateSetListener? = null
+
+    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
+        val c = Calendar.getInstance()
+        val year = c.get(Calendar.YEAR)
+        val month = c.get(Calendar.MONTH)
+        val day = c.get(Calendar.DAY_OF_MONTH)
+
+        return DatePickerDialog(this.activity!!, listener, year, month, day)
+    }
+
+    companion object {
+        fun newInstance(listener: DatePickerDialog.OnDateSetListener): DatePickerFragment {
+            val fragment = DatePickerFragment()
+            fragment.listener = listener
+            return fragment
+        }
+    }
+
+}
