+++ app/src/main/java/com/example/happyroutine/ui/fragment/DietFragment.kt
@@ -1,21 +1,23 @@
-package com.example.happyroutine
+package com.example.happyroutine.ui.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.happyroutine.R
+
 /**
  * A simple [Fragment] subclass.
  */
-class ExerciseFragment : Fragment() {
+class DietFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_exercise, container, false)
+        return inflater.inflate(R.layout.fragment_diet, container, false)
     }
 
 }
