+++ app/src/main/res/layout/activity_forgot_my_password.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/gradient_background"
     android:textAlignment="center"
-    tools:context=".forgot_my_password">
+    tools:context=".ui.activity.forgot_my_password">
 
     <EditText
         android:id="@+id/editText"
