+++ app/src/main/res/layout/activity_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".settings"
+    tools:context=".ui.activity.settings"
     android:background="@drawable/gradient_background">
 
     <View
