+++ app/src/main/res/layout/activity_social.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".social"
+    tools:context=".ui.activity.social"
     android:background="@drawable/gradient_background">
 
     <LinearLayout
