+++ app/src/main/res/layout/fragment_diet.xml
@@ -6,7 +6,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@drawable/gradient_background"
     android:orientation="vertical"
-    tools:context=".DietFragment">
+    tools:context=".ui.fragment.DietFragment">
 
     <ImageButton
         android:id="@+id/setting_onClick"
