+++ app/src/main/res/layout/fragment_edit_profile.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/gradient_background"
     android:fitsSystemWindows="true"
-    tools:context=".EditProfileFragment">
+    tools:context=".ui.fragment.EditProfileFragment">
 
 
     <LinearLayout
