+++ app/src/main/res/layout/fragment_estadisticas_main.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/gradient_background"
     android:fitsSystemWindows="true"
-    tools:context=".EstadisticasMainFragment">
+    tools:context=".ui.activity.EstadisticasMainFragment">
 
 
     <LinearLayout
