+++ app/src/main/res/layout/fragment_exercise.xml
@@ -6,7 +6,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@drawable/gradient_background"
     android:orientation="vertical"
-    tools:context=".ExerciseFragment">
+    tools:context=".ui.fragment.ExerciseFragment">
 
     <ImageButton
         android:id="@+id/setting_onClick"
