+++ app/src/main/res/layout/fragment_social.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/gradient_background"
-    tools:context=".SocialFragment">
+    tools:context=".ui.fragment.SocialFragment">
 
 
     <LinearLayout
